Allow build with either `texlua build.lua` or `l3build`
Newer installations provide the `l3build` executable, which loads the
local `build.lua` configuration file.  Older installations require the
`l3build` module to be _included in_ the configuration file, and for the
config file to be explicitly fed to `texlua`.

Added a few extra lines to `build.lua` to conditionally load the module
when `l3build` is not being called directly, as in Section 1.5 of the
`l3build` documentation.